+++ internal/crypto/object.go
@@ -2,6 +2,7 @@ package crypto
 
 import (
 	"crypto/ecdsa"
+	"crypto/sha256"
 	"errors"
 	"fmt"
 
@@ -12,7 +13,7 @@ import (
 )
 
 // AuthenticateObject checks whether obj is signed correctly by its owner.
-func AuthenticateObject(obj object.Object) error {
+func AuthenticateObject(obj object.Object, fsChain HistoricN3ScriptRunner) error {
 	sig := obj.Signature()
 	if sig == nil {
 		return errMissingSignature
@@ -20,6 +21,7 @@ func AuthenticateObject(obj object.Object) error {
 
 	var ecdsaPub *ecdsa.PublicKey
 	scheme := sig.Scheme()
+	sessionToken := obj.SessionToken()
 	switch scheme {
 	default:
 		return fmt.Errorf("unsupported scheme %v", scheme)
@@ -28,15 +30,19 @@ func AuthenticateObject(obj object.Object) error {
 		if ecdsaPub, err = decodeECDSAPublicKey(sig.PublicKeyBytes()); err != nil {
 			return schemeError(scheme, fmt.Errorf("decode public key: %w", err))
 		}
+	case neofscrypto.N3:
+		if sessionToken != nil {
+			// https://github.com/nspcc-dev/neofs-api/issues/305#issuecomment-2775087206
+			return fmt.Errorf("%s scheme is not supported for objects created with session", scheme)
+		}
 	}
 
-	sessionToken := obj.SessionToken()
 	if sessionToken != nil {
 		// NOTE: update this place for non-ECDSA schemes
 		if !sessionToken.AssertAuthKey((*neofsecdsa.PublicKey)(ecdsaPub)) { // same format for all ECDSA schemes
 			return errors.New("session token is not for object's signer")
 		}
-		if err := AuthenticateToken(sessionToken); err != nil {
+		if err := AuthenticateToken(sessionToken, fsChain); err != nil {
 			return fmt.Errorf("session token: %w", err)
 		}
 		if sessionToken.Issuer() != obj.Owner() {
@@ -53,6 +59,12 @@ func AuthenticateObject(obj object.Object) error {
 		}
 		if sessionToken == nil && user.NewFromECDSAPublicKey(*ecdsaPub) != obj.Owner() {
 			return errors.New("owner mismatches signature")
+		}
+	case neofscrypto.N3:
+		if err := verifyN3ScriptsAtEpoch(fsChain, obj.CreationEpoch(), obj.Owner().ScriptHash(), sig.Value(), sig.PublicKeyBytes(), func() [sha256.Size]byte {
+			return sha256.Sum256(obj.SignedData())
+		}); err != nil {
+			return err
 		}
 	}