+++ internal/crypto/tokens.go
@@ -1,6 +1,7 @@
 package crypto
 
 import (
+	"crypto/sha256"
 	"errors"
 	"fmt"
 
@@ -11,11 +12,15 @@ import (
 // TODO: https://github.com/nspcc-dev/neofs-node/issues/2795 after API stabilization, move some components to SDK
 
 // AuthenticateToken checks whether t is signed correctly by its issuer.
+//
+// If signature scheme is unsupported, [ErrUnsupportedScheme] returns. It also
+// returns when [neofscrypto.N3] scheme is used but fsChain is not provided.
 func AuthenticateToken[T interface {
 	SignedData() []byte
 	Signature() (neofscrypto.Signature, bool)
 	Issuer() user.ID
-}](token T) error {
+	Iat() uint64
+}](token T, fsChain HistoricN3ScriptRunner) error {
 	issuer := token.Issuer()
 	if issuer.IsZero() {
 		return errors.New("missing issuer")
@@ -26,7 +31,7 @@ func AuthenticateToken[T interface {
 	}
 	switch scheme := sig.Scheme(); scheme {
 	default:
-		return fmt.Errorf("unsupported scheme %v", scheme)
+		return ErrUnsupportedScheme(scheme)
 	case neofscrypto.ECDSA_SHA512, neofscrypto.ECDSA_DETERMINISTIC_SHA256, neofscrypto.ECDSA_WALLETCONNECT:
 		pub, err := decodeECDSAPublicKey(sig.PublicKeyBytes())
 		if err != nil {
@@ -37,6 +42,15 @@ func AuthenticateToken[T interface {
 		}
 		if user.NewFromECDSAPublicKey(*pub) != issuer {
 			return errIssuerMismatch
+		}
+	case neofscrypto.N3:
+		if fsChain == nil {
+			return ErrUnsupportedScheme(neofscrypto.N3)
+		}
+		if err := verifyN3ScriptsAtEpoch(fsChain, token.Iat(), issuer.ScriptHash(), sig.Value(), sig.PublicKeyBytes(), func() [sha256.Size]byte {
+			return sha256.Sum256(token.SignedData())
+		}); err != nil {
+			return err
 		}
 	}
 	return nil