+++ pkg/innerring/processors/container/common.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 
 	icrypto "github.com/nspcc-dev/neofs-node/internal/crypto"
+	"github.com/nspcc-dev/neofs-node/pkg/morph/client"
 	cid "github.com/nspcc-dev/neofs-sdk-go/container/id"
 	"github.com/nspcc-dev/neofs-sdk-go/session"
 	"github.com/nspcc-dev/neofs-sdk-go/user"
@@ -27,9 +28,13 @@ type signatureVerificationData struct {
 
 	verifScript []byte
 	invocScript []byte
-	rfc6979     bool
 
 	signedData []byte
+}
+
+type historicN3ScriptRunner struct {
+	*client.Client
+	NetworkState
 }
 
 // verifySignature is a common method of Container service authentication. Asserts that:
@@ -52,7 +57,10 @@ func (cp *Processor) verifySignature(v signatureVerificationData) error {
 			return fmt.Errorf("decode session token: %w", err)
 		}
 
-		if err = icrypto.AuthenticateToken(&tok); err != nil {
+		if err = icrypto.AuthenticateToken(&tok, historicN3ScriptRunner{
+			Client:       cp.cnrClient.Morph(),
+			NetworkState: cp.netState,
+		}); err != nil {
 			return fmt.Errorf("authenticate session token: %w", err)
 		}
 
@@ -80,10 +88,7 @@ func (cp *Processor) verifySignature(v signatureVerificationData) error {
 		return nil
 	}
 
-	if v.rfc6979 {
-		return icrypto.AuthenticateContainerRequestRFC6979(v.ownerContainer, v.verifScript, v.invocScript, v.signedData)
-	}
-	return icrypto.AuthenticateContainerRequestN3(v.ownerContainer, v.invocScript, v.verifScript, v.signedData, cp.cnrClient.Morph())
+	return icrypto.AuthenticateContainerRequest(v.ownerContainer, v.invocScript, v.verifScript, v.signedData, cp.cnrClient.Morph())
 }
 
 func (cp *Processor) checkTokenLifetime(token session.Container) error {